// src/types/types.ts
export interface JobOffer {
  id: string;
  title: string;
  hourly_pay: string;
  benefits: string;
  city: string;
  image_url: string;
  whatsapp_url: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}